Split MMCIF2Dict tests out of test_PDB.py
